Backend/agents/demand_forecast/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.3
prophet==1.1.5
statsmodels==0.14.1
scikit-learn==1.4.0
mlflow==2.9.2
python-dotenv==1.0.0
joblib==1.3.2
psycopg2-binary==2.9.9
